Move around to correct directories after merge
