+++ pipeline-model-definition/src/main/resources/org/jenkinsci/plugins/pipeline/modeldefinition/config/FolderConfig/config.groovy
@@ -23,4 +23,13 @@
  *
  */
 
-echo "Docker Label is: ${dockerLabel()}"
+package org.jenkinsci.plugins.pipeline.modeldefinition.config.FolderConfig
+
+
+def f = namespace(lib.FormTagLib)
+
+f.section(title:_("Pipeline Model Definition")) {
+    f.entry(field: "dockerLabel", title: _("Docker Label")) {
+        f.textbox()
+    }
+}
